@@ -12,33 +12,26 @@ import (
 	"encoding/base64"
 	"encoding/pem"
 	"fmt"
-	"io/ioutil"
+	"io"
+	"log"
 	"os"
 	"strings"
-	"time"
 
 	"google.golang.org/api/compute/v1"
-	"google.golang.org/api/option"
 	"google.golang.org/api/storage/v1"
 
 	"github.com/hashicorp/hcl/v2/hcldec"
-	"github.com/hashicorp/packer-plugin-googlecompute/builder/googlecompute"
-	"github.com/hashicorp/packer-plugin-sdk/common"
+	"github.com/hashicorp/packer-plugin-googlecompute/lib/common"
+	sdk_common "github.com/hashicorp/packer-plugin-sdk/common"
 	packersdk "github.com/hashicorp/packer-plugin-sdk/packer"
 	"github.com/hashicorp/packer-plugin-sdk/template/config"
 	"github.com/hashicorp/packer-plugin-sdk/template/interpolate"
 )
 
 type Config struct {
-	common.PackerConfig `mapstructure:",squash"`
+	sdk_common.PackerConfig `mapstructure:",squash"`
+	common.Authentication   `mapstructure:",squash"`
 
-	//A temporary OAuth 2.0 access token
-	AccessToken string `mapstructure:"access_token" required:"false"`
-	//The JSON file containing your account credentials.
-	//If specified, the account file will take precedence over any `googlecompute` builder authentication method.
-	AccountFile string `mapstructure:"account_file" required:"false"`
-	// This allows service account impersonation as per the [docs](https://cloud.google.com/iam/docs/impersonating-service-accounts).
-	ImpersonateServiceAccount string `mapstructure:"impersonate_service_account" required:"false"`
 	// The service account scopes for launched importer post-processor instance.
 	// Defaults to:
 	//
@@ -79,8 +72,7 @@ type Config struct {
 	//bucket after the import process has completed. "true" means that we should
 	//leave it in the GCS bucket, "false" means to clean it out. Defaults to
 	//`false`.
-	SkipClean           bool   `mapstructure:"skip_clean"`
-	VaultGCPOauthEngine string `mapstructure:"vault_gcp_oauth_engine"`
+	SkipClean bool `mapstructure:"skip_clean"`
 	//A key used to establish the trust relationship between the platform owner and the firmware. You may only specify one platform key, and it must be a valid X.509 certificate.
 	ImagePlatformKey string `mapstructure:"image_platform_key"`
 	//A key used to establish a trust relationship between the firmware and the OS. You may specify multiple comma-separated keys for this value.
@@ -90,8 +82,7 @@ type Config struct {
 	//A database of certificates that have been revoked and will cause the system to stop booting if a boot file is signed with one of them. You may specify single or multiple comma-separated values for this value.
 	ImageForbiddenSignaturesDB []string `mapstructure:"image_forbidden_signatures_db"`
 
-	account *googlecompute.ServiceAccount
-	ctx     interpolate.Context
+	ctx interpolate.Context
 }
 
 type PostProcessor struct {
@@ -141,20 +132,12 @@ func (p *PostProcessor) Configure(raws ...interface{}) error {
 		}
 	}
 
-	if p.config.AccountFile != "" {
-		if p.config.VaultGCPOauthEngine != "" && p.config.ImpersonateServiceAccount != "" {
-			errs = packersdk.MultiErrorAppend(errs, fmt.Errorf("You cannot "+
-				"specify impersonate_service_account, account_file and vault_gcp_oauth_engine at the same time"))
-		}
-		if p.config.AccessToken != "" {
-			errs = packersdk.MultiErrorAppend(errs, fmt.Errorf("You cannot "+
-				"specify access_token and account_file at the same time"))
-		}
-		cfg, err := googlecompute.ProcessAccountFile(p.config.AccountFile)
-		if err != nil {
-			errs = packersdk.MultiErrorAppend(errs, err)
-		}
-		p.config.account = cfg
+	warns, err := p.config.Authentication.Prepare()
+	if err != nil {
+		errs = packersdk.MultiErrorAppend(errs, err)
+	}
+	for _, warn := range warns {
+		log.Printf("[WARN] - %s", warn)
 	}
 
 	if len(p.config.Scopes) == 0 {
@@ -190,8 +173,13 @@ func (p *PostProcessor) PostProcess(ctx context.Context, ui packersdk.Ui, artifa
 	}
 	p.config.ctx.Data = generatedData
 	var err error
-	var opts []option.ClientOption
-	opts, err = googlecompute.NewClientOptionGoogle(p.config.account, p.config.VaultGCPOauthEngine, p.config.ImpersonateServiceAccount, p.config.AccessToken, p.config.Scopes)
+
+	cfg := &common.GCEDriverConfig{
+		Ui:     ui,
+		Scopes: p.config.Scopes,
+	}
+	p.config.Authentication.ApplyDriverConfig(cfg)
+	driver, err := common.NewDriverGCE(*cfg)
 	if err != nil {
 		return nil, false, false, err
 	}
@@ -213,7 +201,12 @@ func (p *PostProcessor) PostProcess(ctx context.Context, ui packersdk.Ui, artifa
 		return nil, false, false, fmt.Errorf("Error rendering gcs_object_name template: %s", err)
 	}
 
-	rawImageGcsPath, err := UploadToBucket(opts, ui, artifact, p.config.Bucket, p.config.GCSObjectName)
+	tarball, err := p.findTarballFromArtifact(artifact)
+	if err != nil {
+		return nil, false, false, err
+	}
+
+	rawImageGcsPath, err := driver.UploadToBucket(p.config.Bucket, p.config.GCSObjectName, tarball)
 	if err != nil {
 		return nil, false, false, err
 	}
@@ -223,23 +216,73 @@ func (p *PostProcessor) PostProcess(ctx context.Context, ui packersdk.Ui, artifa
 		return nil, false, false, err
 	}
 
-	gceImageArtifact, err := CreateGceImage(opts, ui, p.config.ProjectId, rawImageGcsPath, p.config.ImageName, p.config.ImageDescription, p.config.ImageFamily, p.config.ImageLabels, p.config.ImageGuestOsFeatures, shieldedVMStateConfig, p.config.ImageStorageLocations, p.config.ImageArchitecture)
+	var retArtifact *Artifact
+	var retErr error
+
+	imageFeatures := make([]*compute.GuestOsFeature, 0, len(p.config.ImageGuestOsFeatures))
+	for _, v := range p.config.ImageGuestOsFeatures {
+		imageFeatures = append(imageFeatures, &compute.GuestOsFeature{
+			Type: v,
+		})
+	}
+	imageSpec := &compute.Image{
+		Architecture:                 p.config.ImageArchitecture,
+		Description:                  p.config.ImageDescription,
+		Family:                       p.config.ImageFamily,
+		GuestOsFeatures:              imageFeatures,
+		Labels:                       p.config.ImageLabels,
+		Name:                         p.config.ImageName,
+		RawDisk:                      &compute.ImageRawDisk{Source: rawImageGcsPath},
+		SourceType:                   "RAW",
+		ShieldedInstanceInitialState: shieldedVMStateConfig,
+		StorageLocations:             p.config.ImageStorageLocations,
+	}
+
+	imageCh, errCh := driver.CreateImage(p.config.ProjectId, imageSpec)
+	select {
+	case img := <-imageCh:
+		retArtifact = &Artifact{
+			paths: []string{
+				img.SelfLink,
+			},
+		}
+	case err := <-errCh:
+		retErr = err
+	}
+
 	if err != nil {
-		return nil, false, false, err
+		ui.Say(fmt.Sprintf("failed to create image from raw disk: %s", err))
 	}
 
 	if !p.config.SkipClean {
-		err = DeleteFromBucket(opts, ui, p.config.Bucket, p.config.GCSObjectName)
+		ui.Say(fmt.Sprintf("deleting %s from bucket %s", p.config.GCSObjectName, p.config.Bucket))
+		err = driver.DeleteFromBucket(p.config.Bucket, p.config.GCSObjectName)
 		if err != nil {
 			return nil, false, false, err
 		}
 	}
 
-	return gceImageArtifact, false, false, nil
+	return retArtifact, false, false, retErr
+}
+
+func (p PostProcessor) findTarballFromArtifact(artifact packersdk.Artifact) (io.Reader, error) {
+	source := ""
+	for _, path := range artifact.Files() {
+		if strings.HasSuffix(path, ".tar.gz") {
+			source = path
+			break
+		}
+	}
+
+	if source == "" {
+		return nil, fmt.Errorf("No tar.gz file found in list of artifacts")
+	}
+
+	return os.Open(source)
 }
 
 func FillFileContentBuffer(certOrKeyFile string) (*compute.FileContentBuffer, error) {
-	data, err := ioutil.ReadFile(certOrKeyFile)
+	data, err := os.ReadFile(certOrKeyFile)
 	if err != nil {
 		err := fmt.Errorf("Unable to read Certificate or Key file %s", certOrKeyFile)
 		return nil, err
@@ -298,115 +341,4 @@ func CreateShieldedVMStateConfig(imageGuestOsFeatures []string, imagePlatformKey
 		}
 	}
 	return shieldedVMStateConfig, nil
-}
-
-func UploadToBucket(opts []option.ClientOption, ui packersdk.Ui, artifact packersdk.Artifact, bucket string, gcsObjectName string) (string, error) {
-	service, err := storage.NewService(context.TODO(), opts...)
-	if err != nil {
-		return "", err
-	}
-
-	ui.Say("Looking for tar.gz file in list of artifacts...")
-	source := ""
-	for _, path := range artifact.Files() {
-		ui.Say(fmt.Sprintf("Found artifact %v...", path))
-		if strings.HasSuffix(path, ".tar.gz") {
-			source = path
-			break
-		}
-	}
-
-	if source == "" {
-		return "", fmt.Errorf("No tar.gz file found in list of artifacts")
-	}
-
-	artifactFile, err := os.Open(source)
-	if err != nil {
-		err := fmt.Errorf("error opening %v", source)
-		return "", err
-	}
-
-	ui.Say(fmt.Sprintf("Uploading file %v to GCS bucket %v/%v...", source, bucket, gcsObjectName))
-	storageObject, err := service.Objects.Insert(bucket, &storage.Object{Name: gcsObjectName}).Media(artifactFile).Do()
-	if err != nil {
-		ui.Say(fmt.Sprintf("Failed to upload: %v", storageObject))
-		return "", err
-	}
-
-	return storageObject.SelfLink, nil
-}
-
-func CreateGceImage(opts []option.ClientOption, ui packersdk.Ui, project string, rawImageURL string, imageName string, imageDescription string, imageFamily string, imageLabels map[string]string, imageGuestOsFeatures []string, shieldedVMStateConfig *compute.InitialStateConfig, imageStorageLocations []string, imageArchitecture string) (packersdk.Artifact, error) {
-	service, err := compute.NewService(context.TODO(), opts...)
-
-	if err != nil {
-		return nil, err
-	}
-
-	// Build up the imageFeatures
-	imageFeatures := make([]*compute.GuestOsFeature, len(imageGuestOsFeatures))
-	for _, v := range imageGuestOsFeatures {
-		imageFeatures = append(imageFeatures, &compute.GuestOsFeature{
-			Type: v,
-		})
-	}
-
-	gceImage := &compute.Image{
-		Architecture:                 imageArchitecture,
-		Description:                  imageDescription,
-		Family:                       imageFamily,
-		GuestOsFeatures:              imageFeatures,
-		Labels:                       imageLabels,
-		Name:                         imageName,
-		RawDisk:                      &compute.ImageRawDisk{Source: rawImageURL},
-		SourceType:                   "RAW",
-		ShieldedInstanceInitialState: shieldedVMStateConfig,
-		StorageLocations:             imageStorageLocations,
-	}
-
-	ui.Say(fmt.Sprintf("Creating GCE image %v...", imageName))
-	op, err := service.Images.Insert(project, gceImage).Do()
-	if err != nil {
-		ui.Say("Error creating GCE image")
-		return nil, err
-	}
-
-	ui.Say("Waiting for GCE image creation operation to complete...")
-	for op.Status != "DONE" {
-		op, err = service.GlobalOperations.Get(project, op.Name).Do()
-		if err != nil {
-			return nil, err
-		}
-
-		time.Sleep(5 * time.Second)
-	}
-
-	// fail if image creation operation has an error
-	if op.Error != nil {
-		var imageError string
-		for _, error := range op.Error.Errors {
-			imageError += error.Message
-		}
-		err = fmt.Errorf("failed to create GCE image %s: %s", imageName, imageError)
-		return nil, err
-	}
-
-	return &Artifact{paths: []string{op.TargetLink}}, nil
-}
-
-func DeleteFromBucket(opts []option.ClientOption, ui packersdk.Ui, bucket string, gcsObjectName string) error {
-	service, err := storage.NewService(context.TODO(), opts...)
-
-	if err != nil {
-		return err
-	}
-
-	ui.Say(fmt.Sprintf("Deleting import source from GCS %s/%s...", bucket, gcsObjectName))
-	err = service.Objects.Delete(bucket, gcsObjectName).Do()
-	if err != nil {
-		ui.Say(fmt.Sprintf("Failed to delete: %v/%v", bucket, gcsObjectName))
-		return err
-	}
-
-	return nil
 }
