@@ -11,6 +11,7 @@ import (
 	"log"
 
 	"github.com/hashicorp/hcl/v2/hcldec"
+	"github.com/hashicorp/packer-plugin-googlecompute/lib/common"
 	"github.com/hashicorp/packer-plugin-sdk/communicator"
 	"github.com/hashicorp/packer-plugin-sdk/multistep"
 	"github.com/hashicorp/packer-plugin-sdk/multistep/commonsteps"
@@ -46,17 +47,13 @@ func (b *Builder) Prepare(raws ...interface{}) ([]string, []string, error) {
 // Run executes a googlecompute Packer build and returns a packersdk.Artifact
 // representing a GCE machine image.
 func (b *Builder) Run(ctx context.Context, ui packersdk.Ui, hook packersdk.Hook) (packersdk.Artifact, error) {
-	cfg := GCEDriverConfig{
-		Ui:                            ui,
-		ProjectId:                     b.config.ProjectId,
-		Account:                       b.config.account,
-		AccessToken:                   b.config.AccessToken,
-		ImpersonateServiceAccountName: b.config.ImpersonateServiceAccount,
-		Scopes:                        b.config.Scopes,
-		VaultOauthEngineName:          b.config.VaultGCPOauthEngine,
+	cfg := &common.GCEDriverConfig{
+		Ui:        ui,
+		ProjectId: b.config.ProjectId,
 	}
+	b.config.Authentication.ApplyDriverConfig(cfg)
 
-	driver, err := NewDriverGCE(cfg)
+	driver, err := common.NewDriverGCE(*cfg)
 	if err != nil {
 		return nil, err
 	}
@@ -136,12 +133,10 @@ func (b *Builder) Run(ctx context.Context, ui packersdk.Ui, hook packersdk.Hook)
 	}
 
 	artifact := &Artifact{
-		image:     state.Get("image").(*Image),
+		image:     state.Get("image").(*common.Image),
 		driver:    driver,
 		config:    &b.config,
 		StateData: map[string]interface{}{"generated_data": state.Get("generated_data")},
 	}
 	return artifact, nil
 }
-
-// Cancel.
