@@ -14,7 +14,9 @@ import (
 	"os/exec"
 	"strings"
 	"testing"
+	"time"
 
+	"github.com/hashicorp/packer-plugin-googlecompute/lib/common"
 	"github.com/hashicorp/packer-plugin-sdk/acctest"
 )
 
@@ -22,6 +24,8 @@ import (
 var testDataFs embed.FS
 
 func TestAccBuilder_Basic(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/basic.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file %s", err)
@@ -42,6 +46,8 @@ func TestAccBuilder_Basic(t *testing.T) {
 }
 
 func TestAccBuilder_DefaultTokenSource(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/oslogin/default-token.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file %s", err)
@@ -88,6 +94,8 @@ func generateSSHPrivateKey() (string, error) {
 }
 
 func TestAccBuilder_DefaultTokenSourceWithPrivateKey(t *testing.T) {
+	t.Parallel()
+
 	keyFile, err := generateSSHPrivateKey()
 	if err != nil {
 		t.Fatalf("failed to generate SSH private key: %s", err)
@@ -101,7 +109,7 @@ func TestAccBuilder_DefaultTokenSourceWithPrivateKey(t *testing.T) {
 	}
 
 	testCase := &acctest.PluginTestCase{
-		Name:     "googlecompute-packer-default-ts",
+		Name:     "googlecompute-packer-default-ts-with-pkey",
 		Template: fmt.Sprintf(string(tmpl), keyFile),
 		Check: func(buildCommand *exec.Cmd, logfile string) error {
 			if buildCommand.ProcessState != nil {
@@ -132,6 +140,8 @@ func TestAccBuilder_DefaultTokenSourceWithPrivateKey(t *testing.T) {
 }
 
 func TestAccBuilder_WrappedStartupScriptSuccess(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/wrapped-startup-scripts/successful.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file %s", err)
@@ -152,6 +162,8 @@ func TestAccBuilder_WrappedStartupScriptSuccess(t *testing.T) {
 }
 
 func TestAccBuilder_WrappedStartupScriptError(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/wrapped-startup-scripts/errored.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file %s", err)
@@ -172,6 +184,8 @@ func TestAccBuilder_WrappedStartupScriptError(t *testing.T) {
 }
 
 func TestAccBuilder_WithExtraScratchDisk(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/extra_scratch_disk.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file %s", err)
@@ -193,6 +207,8 @@ func TestAccBuilder_WithExtraScratchDisk(t *testing.T) {
 }
 
 func TestAccBuilder_WithExtraPersistentDisk(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/extra_persistent_disk.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file %s", err)
@@ -224,6 +240,8 @@ func TestAccBuilder_WithExtraPersistentDisk(t *testing.T) {
 }
 
 func TestAccBuilder_WithExtraPersistentDiskAndRegions(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/extra_persistent_disk_and_regions.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file %s", err)
@@ -255,6 +273,8 @@ func TestAccBuilder_WithExtraPersistentDiskAndRegions(t *testing.T) {
 }
 
 func TestAccBuilder_WithMultipleDisks(t *testing.T) {
+	t.Parallel()
+
 	tmpl, err := testDataFs.ReadFile("testdata/multiple_disks.pkr.hcl")
 	if err != nil {
 		t.Fatalf("failed to read testdata file: %s", err)
@@ -283,4 +303,107 @@ func TestAccBuilder_WithMultipleDisks(t *testing.T) {
 		},
 	}
 	acctest.TestPlugin(t, testCase)
+}
+
+func TestAccBuilder_ImageArchVariations(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name            string
+		baseImageFamily string
+		arch            string
+		instanceType    string
+		expectedArch    string
+	}{
+		{
+			name:            "googlecompute-packer-with-x86-64-explicit-image-arch",
+			baseImageFamily: "fedora-cloud-38",
+			arch:            "x86_64",
+			instanceType:    "e2-standard-2",
+			expectedArch:    "x86_64",
+		},
+		{
+			name:            "googlecompute-packer-with-x86-64-implicit-image-arch",
+			baseImageFamily: "fedora-cloud-38",
+			arch:            "",
+			instanceType:    "e2-standard-2",
+			// Empty for this case as the source image has an empty
+			// architecture, so this gets forwarded as-is
+			expectedArch: "",
+		},
+		{
+			name:            "googlecompute-packer-with-arm64-explicit-image-arch",
+			baseImageFamily: "ubuntu-minimal-2204-lts-arm64",
+			arch:            "arm64",
+			instanceType:    "t2a-standard-2",
+			expectedArch:    "arm64",
+		},
+		{
+			name:            "googlecompute-packer-with-arm64-implicit-image-arch",
+			baseImageFamily: "ubuntu-minimal-2204-lts-arm64",
+			arch:            "",
+			instanceType:    "t2a-standard-2",
+			expectedArch:    "arm64",
+		},
+	}
+
+	for _, tt := range tests {
+		testRun := tt
+		t.Run(testRun.name, func(t *testing.T) {
+			t.Parallel()
+
+			imageName := fmt.Sprintf("%s-%d", testRun.name, time.Now().UTC().Unix())
+
+			tmpl, err := testDataFs.ReadFile("testdata/image_arch_builds.pkr.hcl")
+			if err != nil {
+				t.Fatalf("failed to read testdata file: %s", err)
+			}
+
+			rawTemplate := fmt.Sprintf(string(tmpl), imageName, testRun.baseImageFamily, testRun.arch, testRun.instanceType)
+
+			testCase := &acctest.PluginTestCase{
+				Name:     testRun.name,
+				Template: rawTemplate,
+				Teardown: func() error {
+					driver, err := common.NewDriverGCE(common.GCEDriverConfig{})
+					if err != nil {
+						return fmt.Errorf("failed to create GCE driver: %s", err)
+					}
+
+					chErr := driver.DeleteImage(os.Getenv("GOOGLE_PROJECT_ID"), imageName)
+					for err := range chErr {
+						return err
+					}
+					return nil
+				},
+				Check: func(buildCommand *exec.Cmd, logfile string) error {
+					if buildCommand.ProcessState != nil {
+						if buildCommand.ProcessState.ExitCode() != 0 {
+							return fmt.Errorf("Bad exit code. Logfile: %s", logfile)
+						}
+					}
+
+					driver, err := common.NewDriverGCE(common.GCEDriverConfig{})
+					if err != nil {
+						return fmt.Errorf("failed to create GCE driver: %s", err)
+					}
+
+					img, err := driver.GetImageFromProject(os.Getenv("GOOGLE_PROJECT_ID"), imageName, false)
+					if err != nil {
+						return fmt.Errorf("failed to get image: %s", err)
+					}
+
+					// Manually uppercase the arch as it cannot be used
+					// for the image name otherwise
+					upperArch := strings.ToUpper(testRun.expectedArch)
+					if img.Architecture != upperArch {
+						return fmt.Errorf("image architecture mismatch, expected %q, got %q", upperArch, img.Architecture)
+					}
+
+					return nil
+				},
+			}
+			acctest.TestPlugin(t, testCase)
+		})
+	}
 }
