@@ -9,8 +9,10 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/hashicorp/packer-plugin-googlecompute/lib/common"
 	"github.com/hashicorp/packer-plugin-sdk/multistep"
 	packersdk "github.com/hashicorp/packer-plugin-sdk/packer"
+	"google.golang.org/api/compute/v1"
 )
 
 // StepCreateImage represents a Packer build step that creates GCE machine
@@ -23,7 +25,7 @@ type StepCreateImage int
 // instance must be deleted and the disk retained before doing this step.
 func (s *StepCreateImage) Run(ctx context.Context, state multistep.StateBag) multistep.StepAction {
 	config := state.Get("config").(*Config)
-	driver := state.Get("driver").(Driver)
+	driver := state.Get("driver").(common.Driver)
 	ui := state.Get("ui").(packersdk.Ui)
 
 	if config.SkipCreateImage {
@@ -46,10 +48,28 @@ func (s *StepCreateImage) Run(ctx context.Context, state multistep.StateBag) mul
 
 	ui.Say("Creating image...")
 
-	imageCh, errCh := driver.CreateImage(
-		config.ImageProjectId, config.ImageName, config.ImageDescription, config.ImageFamily, config.Zone,
-		config.DiskName, config.ImageLabels, config.ImageLicenses, config.ImageGuestOsFeatures,
-		config.ImageEncryptionKey.ComputeType(), config.ImageStorageLocations)
+	sourceDiskURI := fmt.Sprintf("/compute/v1/projects/%s/zones/%s/disks/%s", config.ProjectId, config.Zone, config.imageSourceDisk)
+
+	imageFeatures := make([]*compute.GuestOsFeature, 0, len(config.ImageGuestOsFeatures))
+	for _, v := range config.ImageGuestOsFeatures {
+		imageFeatures = append(imageFeatures, &compute.GuestOsFeature{
+			Type: v,
+		})
+	}
+	imagePayload := &compute.Image{
+		Architecture:       config.ImageArchitecture,
+		Description:        config.ImageDescription,
+		Name:               config.ImageName,
+		Family:             config.ImageFamily,
+		Labels:             config.ImageLabels,
+		Licenses:           config.ImageLicenses,
+		GuestOsFeatures:    imageFeatures,
+		ImageEncryptionKey: config.ImageEncryptionKey.ComputeType(),
+		SourceDisk:         sourceDiskURI,
+		SourceType:         "RAW",
+		StorageLocations:   config.ImageStorageLocations,
+	}
+	imageCh, errCh := driver.CreateImage(config.ImageProjectId, imagePayload)
 	var err error
 	select {
 	case err = <-errCh:
