@@ -1,6 +1,6 @@
 // Code generated by "packer-sdc mapstructure-to-hcl2"; DO NOT EDIT.
 
-package googlecompute
+package common
 
 import (
 	"github.com/hashicorp/hcl/v2/hcldec"
@@ -11,6 +11,7 @@ import (
 // Where the contents of a field with a `mapstructure:,squash` tag are bubbled up.
 type FlatBlockDevice struct {
 	AttachmentMode    *string                    `mapstructure:"attachment_mode" cty:"attachment_mode" hcl:"attachment_mode"`
+	CreateImage       *bool                      `mapstructure:"create_image" cty:"create_image" hcl:"create_image"`
 	DeviceName        *string                    `mapstructure:"device_name" cty:"device_name" hcl:"device_name"`
 	DiskEncryptionKey *FlatCustomerEncryptionKey `mapstructure:"disk_encryption_key" cty:"disk_encryption_key" hcl:"disk_encryption_key"`
 	DiskName          *string                    `mapstructure:"disk_name" cty:"disk_name" hcl:"disk_name"`
@@ -22,6 +23,7 @@ type FlatBlockDevice struct {
 	VolumeSize        *int                       `mapstructure:"volume_size" required:"true" cty:"volume_size" hcl:"volume_size"`
 	VolumeType        *BlockDeviceType           `mapstructure:"volume_type" required:"true" cty:"volume_type" hcl:"volume_type"`
 	SourceImage       *string                    `mapstructure:"source_image" required:"false" cty:"source_image" hcl:"source_image"`
+	Zone              *string                    `mapstructure:"_" cty:"_" hcl:"_"`
 }
 
 // FlatMapstructure returns a new FlatBlockDevice.
@@ -37,6 +39,7 @@ func (*BlockDevice) FlatMapstructure() interface{ HCL2Spec() map[string]hcldec.S
 func (*FlatBlockDevice) HCL2Spec() map[string]hcldec.Spec {
 	s := map[string]hcldec.Spec{
 		"attachment_mode":     &hcldec.AttrSpec{Name: "attachment_mode", Type: cty.String, Required: false},
+		"create_image":        &hcldec.AttrSpec{Name: "create_image", Type: cty.Bool, Required: false},
 		"device_name":         &hcldec.AttrSpec{Name: "device_name", Type: cty.String, Required: false},
 		"disk_encryption_key": &hcldec.BlockSpec{TypeName: "disk_encryption_key", Nested: hcldec.ObjectSpec((*FlatCustomerEncryptionKey)(nil).HCL2Spec())},
 		"disk_name":           &hcldec.AttrSpec{Name: "disk_name", Type: cty.String, Required: false},
@@ -48,6 +51,7 @@ func (*FlatBlockDevice) HCL2Spec() map[string]hcldec.Spec {
 		"volume_size":         &hcldec.AttrSpec{Name: "volume_size", Type: cty.Number, Required: false},
 		"volume_type":         &hcldec.AttrSpec{Name: "volume_type", Type: cty.String, Required: false},
 		"source_image":        &hcldec.AttrSpec{Name: "source_image", Type: cty.String, Required: false},
+		"_":                   &hcldec.AttrSpec{Name: "_", Type: cty.String, Required: false},
 	}
 	return s
 }
